// src/interfaces/LocationInterfaces.tsx
export interface Location {
  lat: number;
  lng: number;
}

export interface Address extends Location {
  address: string;
  viewport?: Viewport;
}

export interface GeocodeResults {
  location: Location;
  viewport: Viewport;
}

export interface Viewport {
  northeast: Location;
  southwest: Location;
}

// Generated by https://quicktype.io

export interface AutocompleteResult {
  description: string;
  distance_meters: number;
  matched_substrings: MatchedSubstring[];
  place_id: string;
  reference: string;
  structured_formatting: StructuredFormatting;
  terms: Term[];
  types: string[];
}

export interface MatchedSubstring {
  length: number;
  offset: number;
}

export interface StructuredFormatting {
  main_text: string;
  main_text_matched_substrings: MatchedSubstring[];
  secondary_text: string;
}

export interface Term {
  offset: number;
  value: string;
}